requirements.txt
accelerate==1.6.0
aiofiles==24.1.0
aiohappyeyeballs==2.4.6
aiohttp==3.11.13
aiosignal==1.3.2
airportsdata==20250224
altair==5.5.0
altgraph==0.17.4
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
astor==0.8.1
attrs==25.1.0
auto_gptq==0.7.1
backoff==2.2.1
bcrypt==4.3.0
beautifulsoup4==4.13.3
bitsandbytes==0.45.5
blake3==1.0.4
blinker==1.9.0
build==1.2.2.post1
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
chroma-hnswlib==0.7.6
chromadb==1.0.4
click==8.1.8
cloudpickle==3.1.1
cmake==4.0.0
colorama==0.4.6
coloredlogs==15.0.1
compressed-tensors==0.9.2
contourpy==1.3.1
cycler==0.12.1
dataclasses-json==0.6.7
datasets==3.3.2
Deprecated==1.2.18
depyf==0.18.0
diffusers==0.24.0
dill==0.3.8
diskcache==5.6.3
distro==1.9.0
dnspython==2.7.0
durationpy==0.9
einops==0.8.1
email_validator==2.2.0
et_xmlfile==2.0.0
faiss-cpu==1.10.0
fastapi==0.115.9
fastapi-cli==0.0.7
fdb==2.0.2
ffmpy==0.5.0
filelock==3.18.0
flatbuffers==25.2.10
fonttools==4.57.0
frozenlist==1.5.0
future==1.0.0
gekko==1.2.1
gguf==0.10.0
gitdb==4.0.12
GitPython==3.1.44
google-auth==2.38.0
googleapis-common-protos==1.69.2
gradio==5.24.0
gradio_client==1.8.0
greenlet==3.1.1
groovy==0.1.2
grpcio==1.71.0
h11==0.14.0
# httpx==0.27.2
httpx-sse==0.4.0
huggingface-hub==0.30.2
humanfriendly==10.0
idna==3.10
importlib_metadata==8.4.0
importlib_resources==6.5.2
iniconfig==2.1.0
interegular==0.3.3
Jinja2==3.1.6
jiter==0.9.0
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.8
kubernetes==32.0.1
langchain==0.3.23
langchain-community==0.3.21
langchain-core==0.3.51
langchain-text-splitters==0.3.8
langsmith==0.1.147
lark==1.2.2
lm-format-enforcer==0.10.11
loralib==0.1.2
markdown-it-py==3.0.0
markdownify==1.1.0
MarkupSafe==2.1.5
marshmallow==3.26.1
matplotlib==3.10.1
mcp==1.6.0
mcpdoc==0.0.8
mdurl==0.1.2
mistral_common==1.5.4
mmh3==5.1.0
monotonic==1.6
mpmath==1.3.0
msgspec==0.19.0
multidict==6.1.0
multiprocess==0.70.16
mypy-extensions==1.0.0
narwhals==1.34.0
nest-asyncio==1.6.0
networkx==3.3
ninja==1.11.1.4
nltk==3.9.1
numpy==1.26.4
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-runtime-cu12==12.8.90
oauthlib==3.2.2
# ollama==0.3.3
onnxruntime==1.21.0
openai==1.72.0
opencv-python==4.8.1.78
opencv-python-headless==4.11.0.86
openpyxl==3.1.5
opentelemetry-api==1.27.0
opentelemetry-exporter-otlp-proto-common==1.27.0
opentelemetry-exporter-otlp-proto-grpc==1.27.0
opentelemetry-instrumentation==0.48b0
opentelemetry-instrumentation-asgi==0.48b0
opentelemetry-instrumentation-fastapi==0.48b0
opentelemetry-proto==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-semantic-conventions==0.48b0
opentelemetry-util-http==0.48b0
orjson==3.10.16
outlines==0.1.11
outlines_core==0.1.26
overrides==7.7.0
packaging==23.2
pandas==2.2.1
partial-json-parser==0.2.1.1.post5
pefile==2023.2.7
peft
pillow==10.4.0
pluggy==1.5.0
posthog==3.24.0
prometheus-fastapi-instrumentator==7.1.0
prometheus_client==0.21.1
propcache==0.3.0
protobuf==4.25.6
psutil==7.0.0
py-cpuinfo==9.0.0
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycountry==24.6.1
pydantic==2.11.3
pydantic-settings==2.8.1
pydantic_core==2.33.1
pydeck==0.9.1
pydub==0.25.1
Pygments==2.19.1
pyinstaller==6.12.0
pyinstaller-hooks-contrib==2025.2
pyodbc==5.0.1
pyparsing==3.2.3
PyPika==0.48.9
pyproject_hooks==1.2.0
pyreadline3==3.5.4
pytest==8.3.5
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==3.3.0
python-multipart==0.0.20
pytz==2025.1
pywin32-ctypes==0.2.3
PyYAML==6.0.2
pyzmq==26.4.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==13.9.4
rich-toolkit==0.14.1
rouge==1.0.1
rpds-py==0.24.0
rsa==4.9
ruff==0.11.5
safehttpx==0.1.6
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
semantic-version==2.10.0
sentence-transformers==2.2.2
sentencepiece==0.2.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.29
sse-starlette==2.2.1
starlette==0.45.3
streamlit==1.32.0
sympy==1.13.1
tenacity==8.5.0
threadpoolctl==3.5.0
tiktoken==0.9.0
tokenizers==0.15.2
toml==0.10.2
tomlkit==0.13.2
torch==2.6.0
torchaudio==2.6.0
torchvision==0.21.0
tornado==6.4.2
tqdm==4.67.1
transformers
typer==0.15.2
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.13.1
tzdata==2025.1
urllib3==2.3.0
uv==0.6.13
uvicorn==0.34.0
watchdog==4.0.0
watchfiles==1.0.5
websocket-client==1.8.0
websockets==15.0.1
wrapt==1.17.2
xxhash==3.5.0
yarl==1.18.3
zipp==3.21.0
